[frontend][tests][refactor] export LocationDisplay to its own component, export parseValue and formateDate to inputHelper.js
