+++ frontend/src/__tests__/SignupForm.test.js
@@ -4,9 +4,11 @@ import {
 } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 import axiosMock from 'axios'
-import { MemoryRouter, Route, withRouter } from 'react-router-dom'
+import { MemoryRouter, Route } from 'react-router-dom'
 import user from '../__mocks__/user'
 import SignupForm from '../components/SignupForm'
+import LocationDisplay from '../services/testHelpers/LocationDisplay'
+import inputHelper from '../services/testHelpers/inputHelper'
 
 jest.mock('axios')
 afterAll(cleanup)
@@ -25,26 +27,6 @@ const DATE_OF_BIRTH_HELPERTEXT = 'Birthday must be between 1900 and 2018'
 const SERVER_DOWN = 'couldn\'t connect to the server'
 const LOGIN_PATH = '/login'
 
-/**
- * @param input: TextField input
- * Helper function that wraps input to a target.value = input object
- * */
-const parseValue = (input) => ({
-  target: {
-    value: input,
-  },
-})
-
-/**
- * @param date: Date object
- * @return string as YYYY-MM-DD
- * */
-const formatDate = (date) => {
-  const year = date.getFullYear()
-  const month = (date.getMonth() + 1).toString().padStart(2, '0')
-  const day = date.getDay().toString().padStart(2, '0')
-  return `${year}-${month}-${day}`
-}
 
 describe('SignupForm tests', () => {
   test('shows an error when the server is down', async () => {
@@ -59,7 +41,7 @@ describe('SignupForm tests', () => {
     const unregisteredUser = {
       username: user.username,
       email: user.email,
-      dateOfBirth: formatDate(new Date(user.dateOfBirth)),
+      dateOfBirth: inputHelper.formatDate(new Date(user.dateOfBirth)),
       fullname: user.fullname,
       password: 'PassWord123',
     }
@@ -69,11 +51,11 @@ describe('SignupForm tests', () => {
     })
 
     expect(getByTestId('signup-container')).toBeInTheDocument()
-    fireEvent.change(getByTestId('username-input'), parseValue(unregisteredUser.username))
-    fireEvent.change(getByTestId('email-input'), parseValue(unregisteredUser.email))
-    fireEvent.change(getByTestId('dateOfBirth-input'), parseValue(unregisteredUser.dateOfBirth))
-    fireEvent.change(getByTestId('password-input'), parseValue(unregisteredUser.password))
-    fireEvent.change(getByTestId('fullname-input'), parseValue(unregisteredUser.fullname))
+    fireEvent.change(getByTestId('username-input'), inputHelper.parseValue(unregisteredUser.username))
+    fireEvent.change(getByTestId('email-input'), inputHelper.parseValue(unregisteredUser.email))
+    fireEvent.change(getByTestId('dateOfBirth-input'), inputHelper.parseValue(unregisteredUser.dateOfBirth))
+    fireEvent.change(getByTestId('password-input'), inputHelper.parseValue(unregisteredUser.password))
+    fireEvent.change(getByTestId('fullname-input'), inputHelper.parseValue(unregisteredUser.fullname))
     fireEvent.click(getByTestId('submit-button'))
 
     await waitForElement(() => getByTestId('signup-container'))
@@ -92,7 +74,7 @@ describe('SignupForm tests', () => {
     const unregisteredUser = {
       username: user.username,
       email: user.email,
-      dateOfBirth: formatDate(new Date(user.dateOfBirth)),
+      dateOfBirth: inputHelper.formatDate(new Date(user.dateOfBirth)),
       fullname: user.fullname,
       password: 'PassWord123',
     }
@@ -107,11 +89,11 @@ describe('SignupForm tests', () => {
     })
 
     expect(getByTestId('signup-container')).toBeInTheDocument()
-    fireEvent.change(getByTestId('username-input'), parseValue(unregisteredUser.username))
-    fireEvent.change(getByTestId('email-input'), parseValue(unregisteredUser.email))
-    fireEvent.change(getByTestId('dateOfBirth-input'), parseValue(unregisteredUser.dateOfBirth))
-    fireEvent.change(getByTestId('password-input'), parseValue(unregisteredUser.password))
-    fireEvent.change(getByTestId('fullname-input'), parseValue(unregisteredUser.fullname))
+    fireEvent.change(getByTestId('username-input'), inputHelper.parseValue(unregisteredUser.username))
+    fireEvent.change(getByTestId('email-input'), inputHelper.parseValue(unregisteredUser.email))
+    fireEvent.change(getByTestId('dateOfBirth-input'), inputHelper.parseValue(unregisteredUser.dateOfBirth))
+    fireEvent.change(getByTestId('password-input'), inputHelper.parseValue(unregisteredUser.password))
+    fireEvent.change(getByTestId('fullname-input'), inputHelper.parseValue(unregisteredUser.fullname))
     fireEvent.click(getByTestId('submit-button'))
 
     rerender(
@@ -149,11 +131,11 @@ describe('SignupForm tests', () => {
       password: 'azzzzzzzsf',
     }
 
-    fireEvent.change(getByTestId('username-input'), parseValue(unregisteredUser.username))
-    fireEvent.change(getByTestId('email-input'), parseValue(unregisteredUser.email))
-    fireEvent.change(getByTestId('dateOfBirth-input'), parseValue(unregisteredUser.dateOfBirth))
-    fireEvent.change(getByTestId('password-input'), parseValue(unregisteredUser.password))
-    fireEvent.change(getByTestId('fullname-input'), parseValue(unregisteredUser.fullname))
+    fireEvent.change(getByTestId('username-input'), inputHelper.parseValue(unregisteredUser.username))
+    fireEvent.change(getByTestId('email-input'), inputHelper.parseValue(unregisteredUser.email))
+    fireEvent.change(getByTestId('dateOfBirth-input'), inputHelper.parseValue(unregisteredUser.dateOfBirth))
+    fireEvent.change(getByTestId('password-input'), inputHelper.parseValue(unregisteredUser.password))
+    fireEvent.change(getByTestId('fullname-input'), inputHelper.parseValue(unregisteredUser.fullname))
     fireEvent.click(getByTestId('submit-button'))
     expect(container.querySelector('#fullName-helper-text').textContent).toEqual(FULLNAME_HELPERTEXT)
     expect(container.querySelector('#username-helper-text').textContent).toEqual(USERNAME_HELPERTEXT)
@@ -168,10 +150,6 @@ describe('SignupForm tests', () => {
   })
 
   test('redirects to /login when login link is clicked', async () => {
-    const LocationDisplay = withRouter(({ location }) => (
-      <div data-testid="location-display">{location.pathname}</div>
-    ))
-
     const { getByTestId, queryByTestId } = render(
       <MemoryRouter initialEntries={['/register']}>
         <Route path="/register">
